third_party/WebKit/LayoutTests/fast/js/Promise-native-then.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise native then.');

window.jsTestIsAsync = true;

var resolve;
var promise = new Promise(function(r) { resolve = r; });
var newPromise = window.internals.addOneToPromise(promise);
console.log("Got a new promise: " + newPromise);
newPromise.then(function(presult) {
    result = presult;
    shouldBe('result', '11');
    finishJSTest();
}, function() {
    testFailed("New promise rejected");
    finishJSTest();
});

resolve(10);

</script>
</body>
</html>
