third_party/WebKit/LayoutTests/fast/js/Promise-native-then-expected.txt

CONSOLE MESSAGE: line 17: Got a new promise: [object Promise]
Test Promise native then.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is 11
PASS successfullyParsed is true

TEST COMPLETE
